% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_api.R
\name{fec_api}
\alias{fec_api}
\title{Query the Open FEC API}
\usage{
fec_api(path, api_key, query_params = list())
}
\arguments{
\item{path}{String: The path of the endpoint you would like you query}

\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Query the base Open FEC API with any path or parameters you like
}
\examples{
fec_api("/candidates/", "YOUR_API_KEY_GOES_HERE", query_params = list(q = c("Obama")))
}
\keyword{api}
