% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_endpoints.R
\name{get_candidate}
\alias{get_candidate}
\title{Get Information on a Candidate}
\usage{
get_candidate(candidate_id, api_key, query_params = list())
}
\arguments{
\item{candidate_id}{String: The id for the candidate of interest.}

\item{api_key}{String: Your Open FEC API key.}

\item{query_params}{List of KV Pairs: Additional query paramters to be included.}
}
\description{
Pulls information for a candidated, identified by a candidate id
}
\examples{
get_candidate("P60019015", "YOUR_API_KEY_HERE")
}
\keyword{candidate}
