% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_endpoints.R
\name{get_candidate_committees}
\alias{get_candidate_committees}
\title{Find Committees that Support a Candidate}
\usage{
get_candidate_committees(candidate_id, api_key, query_params = list())
}
\arguments{
\item{candidate_id}{String: The id for the candidate of interest.}

\item{api_key}{String: Your Open FEC API key.}

\item{query_params}{List of KV Pairs: Additional query paramters to be included.}
}
\description{
Fetches a list of committees tied to a given candidate ID
}
\examples{
get_candidate_committees("P60019015", "YOUR_API_KEY_HERE")
}
\keyword{candidate}
