% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_endpoints.R
\name{get_candidate_history}
\alias{get_candidate_history}
\title{Get History a Candidate}
\usage{
get_candidate_history(candidate_id, api_key, query_params = list())
}
\arguments{
\item{candidate_id}{String: The id for the candidate of interest.}

\item{api_key}{String: Your Open FEC API key.}

\item{query_params}{List of KV Pairs: Additional query paramters to be included.}
}
\description{
Pulls information on a candidate's characteristics over time.
}
\examples{
get_candidate_history("P60019015", "YOUR_API_KEY_HERE")
}
\keyword{candidate}
\keyword{history}
