% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_endpoints.R
\name{get_candidates}
\alias{get_candidates}
\title{Get Candidates}
\usage{
get_candidates(api_key, query_params = list())
}
\arguments{
\item{api_key}{String: Your Open FEC API key.}

\item{query_params}{List of KV Pairs: Additional query paramters to be included.}
}
\description{
Fetch basic information about candidates, and use parameters to filter results to the candidates you're looking for.
}
\examples{
get_candidates("YOUR_API_KEY_HERE", query_params = list(state = "IL"))
}
\keyword{candidate}
