% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_endpoints.R
\name{get_candidates_totals}
\alias{get_candidates_totals}
\title{Get Candidate Totals}
\usage{
get_candidates_totals(api_key, query_params = list())
}
\arguments{
\item{api_key}{String: Your Open FEC API key.}

\item{query_params}{List of KV Pairs: Additional query paramters to be included.}
}
\description{
Aggregated candidate receipts and disbursements grouped by cycle.
}
\examples{
get_candidates_totals("YOUR_API_KEY_HERE", query_params = list(state = "IL"))
}
\keyword{candidate}
\keyword{totals}
