% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_endpoints.R
\name{get_committee}
\alias{get_committee}
\title{Get Information on a Committee}
\usage{
get_committee(committee_id, api_key, query_params = list())
}
\arguments{
\item{committee_id}{String: The Open FEC committee id for the committee you hope to query about}

\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Fetches information on a given committee, identified by committee ID
}
\examples{
get_committee("C00381442", "YOUR_API_KEY_HERE")
}
\keyword{committee}
\keyword{financial}
