% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_endpoints.R
\name{get_committee_candidates_history}
\alias{get_committee_candidates_history}
\title{Get History of Candidates in a Committee}
\usage{
get_committee_candidates_history(committee_id, api_key, query_params = list())
}
\arguments{
\item{committee_id}{String: The id for the committee of interest.}

\item{api_key}{String: Your Open FEC API key.}

\item{query_params}{List of KV Pairs: Additional query paramters to be included.}
}
\description{
Find out characteristics over time for candidates in a committee.
}
\examples{
get_committee_candidates_history("P60019015", "YOUR_API_KEY_HERE")
}
\keyword{candidate}
\keyword{committee}
\keyword{history}
