% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_endpoints.R
\name{get_committee_financial_reports}
\alias{get_committee_financial_reports}
\title{Query for Committee Financial reports}
\usage{
get_committee_financial_reports(committee_id, api_key, query_params = list())
}
\arguments{
\item{committee_id}{String: The Open FEC committee id for the committee you hope to query about}

\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Provides financial reports for a given committee id
}
\examples{
get_committee_financial_reports("C00381442", "YOUR_API_KEY_HERE")
}
\keyword{committee}
\keyword{financial}
