% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_endpoints.R
\name{get_committee_history}
\alias{get_committee_history}
\title{Get History of a Committee}
\usage{
get_committee_history(committee_id, api_key, query_params = list())
}
\arguments{
\item{committee_id}{String: The Open FEC committee id for the committee you hope to query about}

\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Pulls information on a committee's characteristics over time.
}
\examples{
get_committee_history("C00381442", "YOUR_API_KEY_HERE")
}
\keyword{committee}
\keyword{financial}
