% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_endpoints.R
\name{get_election_financial_summary}
\alias{get_election_financial_summary}
\title{Query for an Election Financial Summary}
\usage{
get_election_financial_summary(api_key, query_params = list())
}
\arguments{
\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Look at the top-level financial information for all candidates running for the same office.
}
\examples{
get_election_financial_summary("YOUR_API_KEY_HERE", query_params = list(cycle = "2016", office = "senate", state = "IL"))
}
\keyword{election}
\keyword{financial}
\keyword{summary}
