% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings_endpoints.R
\name{get_filings}
\alias{get_filings}
\title{Get FEC Filings based on Query Params}
\usage{
get_filings(api_key, query_params = list())
}
\arguments{
\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Provides FEC Filing records based on query parameters
}
\examples{
get_filings("YOUR_API_KEY_HERE", query_params = list(state = "IL"))
}
\keyword{filings}
