% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings_endpoints.R
\name{get_filings_for_candidate}
\alias{get_filings_for_candidate}
\title{Get FEC Filings for a Candidate}
\usage{
get_filings_for_candidate(candidate_id, api_key, query_params = list())
}
\arguments{
\item{candidate_id}{String: The Open FEC candidate id for the candidate you hope to query about}

\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Provides FEC Filing records for a given candidate id
}
\examples{
get_filings_for_candidate("P60019015", "YOUR_API_KEY_HERE")
}
\keyword{candidate}
\keyword{filings}
