% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings_endpoints.R
\name{get_filings_for_committee}
\alias{get_filings_for_committee}
\title{Get FEC Filings for a Committee}
\usage{
get_filings_for_committee(committee_id, api_key, query_params = list())
}
\arguments{
\item{committee_id}{String: The Open FEC committee id for the committee you hope to query about}

\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Provides FEC Filing records for a given committee id
}
\examples{
get_filings_for_candidate("C00381442", "YOUR_API_KEY_HERE")
}
\keyword{committee}
\keyword{filings}
