% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_endpoints.R
\name{get_financial_totals_by_entity}
\alias{get_financial_totals_by_entity}
\title{Query for financial totals by entity type}
\usage{
get_financial_totals_by_entity(api_key, query_params = list())
}
\arguments{
\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
Provides cumulative receipt totals by entity type, over a two year cycle. Totals are adjusted to avoid double counting.
}
\examples{
get_financial_totals_by_entity("YOUR_API_KEY_HERE", query_params = list(cycle = "2016", state = "IL"))
}
\keyword{entity}
\keyword{financial}
\keyword{total}
