% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_endpoints.R
\name{search_for_election}
\alias{search_for_election}
\title{Search for an Election's Financial Records}
\usage{
search_for_election(api_key, query_params = list())
}
\arguments{
\item{api_key}{String: Your Open FEC API key}

\item{query_params}{List of KV Pairs: Additional query paramaters to be included}
}
\description{
List elections by cycle, office, state, and district.
}
\examples{
search_for_election("YOUR_API_KEY_HERE", query_params = list(cycle = "2016", office = "house", state = "IL", district = "01"))
}
\keyword{election}
\keyword{financial}
\keyword{search}
